Attempt to fix mail delivery problem
The e-mail is configured on each request. But in the rake task there
is no request. Because of this I think it is just usng the Rails
defaults (which is not the right config).

This needs to be refactored as it is just a copy/paste hack but I wanted
to confirm this fixes my issue first.